apps/insights/pyproject.toml
[tool.poetry]
name = "insights"
version = "1.0.0"
description = "SHINJU DATE Insights API for video analytics"
authors = ["SHINJU DATE Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.119.1"
uvicorn = "^0.38.0"
janome = "^0.5.0"
supabase = "^2.22.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.1"
poethepoet = "^0.37.0"

[tool.poe.tasks]
dev = "uvicorn api.index:app --reload --host 0.0.0.0"
format = "ruff format ."
lint = "ruff check ."

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"
